+++ app.js
@@ -1,7 +1,6 @@
 const express = require("express");
 const logger = require("morgan");
 const cors = require("cors");
-
 const contactsRouter = require("./routes/api/contacts");
 
 const app = express();
@@ -11,13 +10,10 @@ const formatsLogger = app.get("env") === "development" ? "dev" : "short";
 app.use(logger(formatsLogger));
 app.use(cors());
 app.use(express.json());
-
 app.use("/api/contacts", contactsRouter);
-
 app.use((req, res) => {
   res.status(404).json({ message: "Not found" });
 });
-
 app.use((err, req, res, next) => {
   const { status = 500, message = "server error" } = err;
   res.status(status).json({ message });
