+++ controllers/contacts/index.js
@@ -1,0 +1,12 @@
+const getAll = require("./getAll");
+const getById = require("./getById");
+const addContacts = require("./addContacts");
+const deleteContacts = require("./deleteContacts");
+const updateContacts = require("./updateContacts");
+module.exports = {
+  getAll,
+  getById,
+  addContacts,
+  deleteContacts,
+  updateContacts,
+};
