+++ routes/api/contacts.js
@@ -1,80 +1,16 @@
 const express = require("express");
-const contacts = require("../../models/contacts");
-const { generationError } = require("../../helpers");
+const ctrl = require("../../controllers/contacts");
 
-const Joi = require("joi");
 const router = express.Router();
-const contactsSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
 
-router.get("/", async (_, res, next) => {
-  try {
-    const result = await contacts.listContacts();
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.get("/", ctrl.getAll);
 
-router.get("/:contactId", async (req, res, next) => {
-  try {
-    const id = req.params.contactId;
-    const result = await contacts.getContactById(id);
+router.get("/:contactId", ctrl.getById);
 
-    if (!result) {
-      throw generationError(404, "Not found");
-    }
-    res.json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.post("/", ctrl.addContacts);
 
-router.post("/", async (req, res, next) => {
-  try {
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      throw generationError(400, "missing required name field");
-    }
-    const result = await contacts.addContact(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.delete("/:contactId", ctrl.deleteContacts);
 
-router.delete("/:contactId", async (req, res, next) => {
-  try {
-    const id = req.params.contactId;
-    const result = await contacts.removeContact(id);
-    if (!result) {
-      throw generationError(404, "Not found");
-    }
-    res.status(200).json({ message: "contact deleted" });
-  } catch (error) {
-    next(error);
-  }
-});
-
-router.put("/:contactId", async (req, res, next) => {
-  try {
-    const id = req.params.contactId;
-    const { error } = contactsSchema.validate(req.body);
-    if (error) {
-      // return res.status(400).json({ message: "missing fields" });
-      throw generationError(400, "missing fields");
-    }
-    const result = await contacts.updateContact(id, req.body);
-    if (!result) {
-      throw generationError(404, "Not found");
-    }
-    res.status(200).json(result);
-  } catch (error) {
-    next(error);
-  }
-});
+router.put("/:contactId", ctrl.updateContacts);
 
 module.exports = router;
