+++ schemes/contactSchema.js
@@ -1,0 +1,7 @@
+const Joi = require("joi");
+const contactsSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+});
+module.exports = contactsSchema;
